requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.0
ollama>=0.1.0
psycopg2-binary>=2.9.9
supabase>=2.0.0
pyttsx3>=2.90
gtts>=2.5.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
pydub>=0.25.1
pygame>=2.5.0
pyaudio>=0.2.14

